third_party/WebKit/LayoutTests/platform/win/fast/text/hyphens/hyphens-parsing-001-expected.txt

This is a testharness.js-based test.
PASS Initial value
FAIL hyphens: autoassert_equals: expected "auto" but got "manual"
PASS hyphens: manual
PASS hyphens: none
FAIL hyphens should inheritassert_equals: expected "auto" but got "manual"
Harness: the test ran to completion.
